refclock_generic: change implicit cast to explcit.
Fix a warning on 32-bit RasPi.

+++ ntpd/refclock_generic.c
@@ -1791,7 +1791,7 @@ local_input(
 	/*
 	 * eat all characters, parsing then and feeding complete samples
 	 */
-	count = rbufp->recv_length;
+	count = (int)rbufp->recv_length;
 	s = (unsigned char *)rbufp->recv_buffer;
 	ts = rbufp->recv_time;
